adding 2-sat problem
