+++ algorithms/problems/input/2sat_20_0.txt
@@ -1,0 +1,21 @@
+20
+-3 -10
+-9 -20
+-5 4
+-9 -9
+-12 2
+-13 -17
+19 2
+-6 -5
+-10 16
+5 19
+12 -10
+-12 8
+12 12
+-16 -16
+3 -19
+10 -8
+-18 -19
+17 -5
+3 20
+-15 -10
