+++ algorithms/problems/input/2sat_2_0.txt
@@ -1,0 +1,3 @@
+2
+2 2
+-2 -2
