+++ algorithms/problems/input/2sat_400_0.txt
@@ -1,0 +1,401 @@
+400
+166 361
+-330 -136
+-264 350
+203 149
+-304 -267
+157 120
+-31 -193
+-35 -332
+-61 -240
+112 372
+-73 -73
+-37 -116
+-104 101
+334 185
+140 130
+166 -33
+-399 -200
+373 -66
+226 45
+-398 -148
+108 -329
+-302 -145
+4 363
+153 121
+-42 -309
+346 -171
+185 258
+-32 -194
+177 -184
+-52 89
+-73 113
+-389 20
+-111 14
+-48 291
+-86 -333
+-171 38
+253 -115
+163 44
+315 -92
+-327 -58
+-252 -236
+299 -142
+130 260
+82 377
+-214 165
+-107 -165
+220 -380
+-54 290
+-233 114
+-171 -171
+21 345
+-391 222
+-166 -276
+-28 399
+-200 -265
+-99 -143
+185 -240
+-306 -35
+-178 106
+-354 -204
+380 -158
+-240 87
+-292 -374
+-37 -15
+-6 -66
+160 -95
+96 322
+8 391
+157 199
+-374 -148
+83 25
+188 -171
+383 43
+303 -281
+276 166
+31 -229
+265 48
+17 -74
+-231 -354
+141 -248
+-189 242
+327 -329
+160 -210
+314 250
+87 261
+-394 178
+352 -69
+-100 -248
+141 -340
+-182 -304
+262 299
+-134 94
+398 323
+15 337
+230 -331
+-357 335
+36 -273
+-97 57
+-180 -117
+360 310
+-113 246
+-112 133
+-241 -325
+333 187
+389 -258
+255 -370
+212 -277
+-71 261
+-361 -260
+-203 163
+-75 -154
+83 291
+85 -94
+-330 -20
+-263 226
+56 -368
+122 358
+269 112
+389 -75
+-63 -380
+-382 -227
+158 45
+284 118
+122 -212
+-73 -344
+319 -335
+133 377
+22 309
+-370 -273
+106 96
+191 -390
+107 92
+26 114
+27 288
+-349 -141
+228 -125
+66 -147
+24 -245
+-349 352
+13 149
+-130 198
+59 -172
+-255 56
+-234 336
+97 138
+-282 345
+-124 -378
+-242 81
+393 -327
+-60 -28
+350 75
+-113 117
+104 -175
+299 324
+242 -397
+-254 -326
+-213 81
+-13 288
+183 -38
+-335 203
+-110 -59
+376 233
+-98 324
+-175 -183
+391 13
+-357 290
+-135 300
+52 -371
+27 -357
+176 -23
+195 268
+212 219
+-345 -380
+173 -90
+-18 -320
+-74 203
+-145 -53
+-157 353
+227 14
+31 -92
+279 64
+-58 -155
+35 289
+-191 -154
+-391 9
+-291 235
+-13 264
+15 -13
+-298 -145
+-296 390
+-163 -108
+311 131
+376 -115
+124 289
+149 108
+-308 -150
+15 48
+108 -121
+73 -212
+-109 396
+164 -181
+-78 -273
+199 -81
+115 83
+-267 250
+-176 56
+41 -336
+193 62
+371 -279
+1 -102
+331 -255
+122 -175
+181 387
+-159 -347
+-256 190
+82 -322
+166 5
+33 154
+-194 -333
+-343 -247
+37 -190
+-7 -180
+291 94
+-298 310
+-135 396
+-283 -161
+306 39
+-107 -152
+-239 224
+339 372
+-341 342
+-105 129
+-204 -235
+60 -231
+271 269
+270 -225
+153 185
+358 -384
+-236 324
+-252 -120
+180 138
+-2 -81
+5 -206
+-197 -143
+293 -64
+351 39
+-147 30
+193 -162
+87 295
+263 -53
+236 -45
+-32 -4
+-319 -179
+-266 -355
+299 276
+-234 208
+-302 358
+217 311
+70 -113
+-199 -314
+228 125
+88 301
+-319 12
+-30 -329
+34 -70
+-366 -189
+378 84
+398 -176
+64 -265
+244 68
+-171 84
+170 97
+202 48
+-341 346
+-226 -155
+219 -167
+-110 -56
+59 167
+-160 370
+316 107
+-300 -195
+-117 271
+-330 -97
+-176 319
+-227 363
+399 276
+-205 162
+-361 295
+345 -57
+175 134
+101 -24
+-180 358
+61 126
+-34 239
+140 -332
+-203 129
+-316 163
+247 269
+-302 -170
+289 -360
+234 400
+232 57
+-220 167
+211 -1
+396 43
+-166 -72
+-326 198
+63 88
+229 -245
+330 120
+-80 -82
+93 -15
+299 397
+-386 -193
+282 73
+-200 -363
+-278 173
+40 72
+-160 -257
+-400 260
+284 -288
+265 -40
+21 -9
+232 -144
+-132 -328
+398 -89
+-199 -340
+349 346
+168 279
+-58 35
+291 149
+-231 -349
+385 -251
+304 -344
+45 -64
+147 90
+154 -55
+-346 214
+290 389
+210 180
+322 30
+88 165
+350 -365
+69 174
+-28 130
+228 -31
+-344 -155
+347 -293
+66 136
+181 216
+-236 256
+307 104
+123 277
+96 334
+-306 307
+234 4
+-9 -291
+-230 391
+127 156
+233 -299
+-243 -343
+225 247
+306 -21
+-252 -189
+-263 -188
+143 150
+124 352
+-353 332
+-5 22
+-7 338
+-338 140
+-311 365
+-396 86
+206 -374
+-22 -148
+85 -393
+165 175
+-144 -99
+94 -51
+-149 36
+-74 192
+273 -178
+-110 347
+341 368
+105 183
+377 -324
+315 330
+-399 -63
+-154 307
+-317 -169
+-326 -142
+83 -54
+269 -44
+-254 -89
+-222 -202
+202 -134
+-348 -46
+235 -165
+-84 -297
+44 -224
