+++ algorithms/problems/input/2sat_40_1.txt
@@ -1,0 +1,41 @@
+40
+27 11
+-40 -34
+29 -31
+-35 17
+-5 -1
+-31 4
+5 36
+-23 20
+-11 2
+-24 -10
+-15 35
+-23 2
+34 -13
+-24 9
+-14 -22
+34 16
+24 21
+-11 27
+-11 -14
+15 -22
+19 22
+-23 -21
+-26 -37
+-40 27
+-5 31
+-21 9
+-21 36
+2 -9
+10 24
+1 5
+-29 -30
+1 -23
+5 -11
+37 -22
+38 17
+16 -10
+25 6
+19 -30
+40 16
+-24 9
