+++ algorithms/problems/input/2sat_4_1.txt
@@ -1,0 +1,5 @@
+4
+1 3
+-1 4
+-3 4
+4 -1
