+++ algorithms/problems/input/2sat_800_1.txt
@@ -1,0 +1,801 @@
+800
+-357 522
+-674 -764
+-267 -210
+751 -250
+139 -338
+444 303
+478 725
+325 247
+-754 457
+293 -241
+356 -225
+-420 586
+-174 -430
+397 170
+278 339
+583 -667
+-450 -452
+430 756
+-389 -734
+-390 746
+-613 493
+-615 742
+34 -419
+364 149
+276 -658
+-777 716
+2 741
+-484 -506
+112 651
+-673 688
+624 172
+419 419
+-305 -247
+-489 74
+-86 31
+-212 -775
+-461 -737
+-48 -575
+-82 -428
+479 475
+376 353
+517 -672
+-639 -372
+184 -761
+216 335
+-623 -369
+577 -49
+744 670
+-452 34
+-619 513
+-525 780
+225 169
+-410 -195
+-374 336
+412 -649
+-110 -157
+72 -330
+426 -22
+-403 425
+727 431
+-409 708
+362 661
+-773 758
+-642 -262
+-267 -408
+-624 204
+-74 -711
+559 783
+-379 -2
+108 251
+391 455
+301 -527
+-549 381
+-600 599
+-209 760
+356 789
+-689 195
+-432 -371
+-141 603
+10 -174
+-368 -322
+150 758
+-140 109
+15 734
+200 357
+235 750
+-408 541
+540 705
+106 -750
+211 689
+-469 312
+44 96
+81 -160
+-32 -240
+110 -641
+-533 348
+-602 -256
+682 -340
+778 -510
+252 292
+786 772
+-523 234
+-165 -523
+-364 -289
+534 234
+559 -554
+-773 -320
+-746 796
+567 -695
+728 -577
+-332 -696
+599 -544
+-183 574
+-688 -25
+-568 582
+-782 764
+423 167
+-464 751
+41 551
+375 -77
+-44 -7
+91 477
+-384 -307
+771 467
+1 -557
+-350 -13
+154 -709
+-114 -441
+205 -15
+468 -22
+-687 -751
+150 779
+-780 -666
+-432 -159
+-478 70
+-87 226
+134 -152
+-286 172
+-566 -59
+434 242
+795 403
+-84 602
+13 -226
+-773 -426
+152 -106
+-686 -134
+-390 -102
+-48 118
+-635 527
+412 -641
+699 772
+329 345
+565 452
+-460 -265
+334 -469
+-730 521
+-367 -713
+370 434
+467 537
+-105 9
+709 -483
+794 734
+-765 -1
+75 357
+-786 -570
+634 -559
+-18 -540
+553 -613
+737 -742
+348 -262
+605 -66
+442 -140
+-448 80
+-464 475
+692 -48
+682 468
+71 473
+762 635
+783 -472
+436 325
+502 -515
+-77 -639
+18 -227
+-50 -735
+-579 -481
+-512 522
+-24 521
+-355 -537
+150 310
+760 264
+633 536
+-111 555
+373 160
+531 -128
+167 443
+-327 -261
+-570 663
+378 772
+-402 -387
+-683 642
+348 -486
+783 80
+575 518
+-56 491
+767 -409
+10 431
+799 85
+297 -790
+315 710
+-750 416
+-529 -728
+243 64
+737 -523
+552 686
+-625 544
+133 -747
+-666 407
+-775 -489
+401 548
+-11 380
+-752 420
+-657 -780
+-197 671
+540 -298
+-472 128
+-565 596
+276 -789
+192 -756
+-214 17
+391 176
+430 168
+530 -419
+-497 -622
+586 155
+-190 -467
+-143 437
+-370 783
+223 -557
+-74 -674
+-128 -197
+367 -124
+574 664
+-649 -517
+-291 -34
+283 381
+35 -144
+349 -352
+-713 -117
+469 -254
+-430 425
+-313 -336
+4 776
+-471 729
+274 636
+-69 750
+345 695
+-460 779
+-316 -718
+739 173
+-145 763
+414 145
+139 -468
+-254 444
+601 150
+697 -223
+680 -1
+599 -561
+465 -167
+363 434
+351 -160
+593 607
+-329 -95
+-766 -24
+798 -269
+-341 454
+-599 572
+548 -305
+-494 680
+455 -85
+-663 -363
+-221 235
+-678 -489
+-369 -598
+-338 -345
+-745 549
+-769 572
+60 141
+-348 -43
+406 -556
+-34 -192
+-627 51
+145 402
+58 144
+729 634
+-377 530
+359 -670
+407 -208
+-138 -544
+529 -789
+410 -492
+-600 -474
+473 116
+224 455
+-293 -569
+49 -687
+-797 -591
+224 563
+112 177
+596 -510
+212 -606
+-531 526
+37 -511
+-374 -691
+107 -10
+56 211
+211 103
+-483 -139
+271 -34
+186 133
+-223 -251
+-775 -751
+414 -695
+272 -618
+-24 446
+-520 142
+344 208
+85 -113
+-632 773
+267 259
+-249 -596
+-251 -651
+-203 -491
+13 674
+472 723
+-395 694
+-506 224
+-56 -737
+717 -85
+482 764
+153 751
+512 -254
+-113 -461
+593 -80
+-442 161
+652 779
+266 460
+-229 447
+-18 -360
+-138 36
+-411 643
+17 -320
+151 -218
+227 -429
+-643 620
+525 654
+117 -598
+483 547
+162 600
+-173 756
+381 -443
+-11 276
+76 -768
+-717 550
+594 323
+429 -257
+740 -682
+505 -317
+790 457
+582 -28
+764 499
+-342 -582
+332 -706
+607 565
+636 526
+493 728
+340 390
+80 345
+-179 -282
+1 -49
+422 309
+678 -350
+-260 112
+210 685
+213 63
+565 500
+187 -474
+-336 -230
+-110 -240
+-301 -543
+335 -83
+-753 484
+-344 112
+42 779
+271 246
+-679 -504
+163 276
+421 -751
+250 606
+201 780
+532 755
+-16 564
+168 -577
+-257 36
+391 3
+-571 154
+723 653
+751 327
+-706 155
+758 -739
+224 -349
+-154 364
+380 -709
+-533 542
+681 -102
+716 381
+224 -191
+166 411
+-263 -560
+694 521
+-355 -361
+-528 591
+-48 702
+131 -403
+303 -255
+364 356
+-525 53
+354 -601
+-588 -411
+689 -388
+241 -290
+530 -390
+425 -126
+-305 784
+162 -419
+-387 -521
+594 197
+-769 -149
+391 -188
+-624 357
+778 405
+728 -412
+-708 -301
+-685 -657
+369 414
+-620 473
+463 628
+246 -242
+213 -742
+-533 -372
+-354 530
+748 -523
+-245 -137
+47 310
+478 -724
+-751 -81
+657 -587
+-522 -474
+640 267
+-294 397
+315 327
+212 96
+-270 -153
+451 -420
+-151 -595
+-321 648
+410 -509
+-363 -332
+-588 380
+41 -346
+-586 269
+-14 36
+430 -478
+620 -91
+414 549
+721 -704
+310 -127
+267 -31
+-436 -626
+414 161
+-648 -292
+218 -276
+-295 99
+589 672
+-596 686
+-538 -414
+366 -351
+382 -48
+112 703
+257 -584
+-377 151
+395 513
+26 -633
+-260 418
+-657 1
+623 78
+-243 -751
+-280 292
+710 -360
+-564 -504
+45 636
+-770 595
+-477 775
+403 -364
+244 -227
+-590 189
+79 -213
+-799 321
+729 537
+-446 -627
+611 -250
+-16 -81
+-725 499
+-139 -611
+-783 159
+105 560
+-564 -692
+256 125
+-512 71
+-206 666
+-29 340
+-212 383
+-346 620
+-772 445
+-60 -478
+200 630
+-759 674
+587 650
+119 -570
+198 -114
+-548 -144
+-296 651
+-562 -381
+268 -709
+-775 -745
+-113 423
+204 355
+387 -536
+-347 616
+750 -9
+-452 133
+-318 75
+-432 43
+667 217
+757 265
+487 117
+566 294
+-586 398
+-45 305
+434 -172
+343 -671
+-159 393
+82 476
+-413 635
+-496 -325
+68 569
+166 -626
+62 -230
+-479 149
+-682 -16
+-609 512
+-416 -317
+124 -298
+355 -101
+-573 -334
+605 -354
+347 -79
+356 -32
+348 354
+793 -695
+131 -766
+-684 653
+-428 662
+578 -290
+-458 555
+289 -455
+-431 -23
+-470 626
+-196 -314
+426 -458
+-452 -544
+-442 -768
+-399 489
+42 -385
+-665 -325
+604 269
+625 414
+164 427
+-206 672
+-165 -614
+780 649
+34 649
+-18 -369
+370 169
+-527 228
+-753 573
+-526 106
+541 -15
+193 -252
+-644 -221
+149 -575
+783 163
+276 -130
+-753 641
+645 499
+-13 -709
+502 -409
+-65 -20
+-519 -563
+410 -716
+-627 -618
+-732 547
+-609 458
+661 555
+693 211
+-152 634
+-780 64
+134 567
+-541 324
+196 303
+215 459
+-769 729
+-101 -107
+664 99
+392 154
+562 224
+491 43
+395 -168
+110 444
+-795 -18
+454 -638
+232 462
+-688 519
+257 -271
+-732 584
+-386 -718
+-223 -535
+708 -402
+-521 25
+-635 126
+-135 263
+-604 448
+308 48
+-420 176
+-88 -701
+334 401
+108 -107
+-9 203
+-469 -102
+-751 672
+86 -549
+-791 -668
+653 570
+-245 -192
+-368 122
+-87 -654
+-648 446
+-176 -62
+583 -290
+-52 -777
+-784 611
+-70 -447
+-492 77
+-776 -268
+-349 763
+426 -561
+731 -152
+-182 -536
+382 -486
+277 -471
+432 742
+-650 365
+375 665
+-477 -174
+-9 455
+-176 -43
+555 619
+-639 57
+512 -72
+-416 -259
+-189 607
+-245 232
+375 -418
+-107 -5
+131 -24
+420 242
+-484 358
+-582 645
+-56 -78
+207 -746
+-124 -223
+-642 -117
+640 -176
+-295 -582
+-535 559
+-687 -352
+650 -190
+-593 527
+-559 -21
+-204 644
+187 258
+185 -753
+625 -293
+-759 500
+122 -575
+-799 -500
+-543 293
+-39 48
+-689 86
+697 613
+285 253
+290 739
+414 -538
+-639 -610
+19 -297
+550 -669
+444 157
+602 284
+703 -389
+193 154
+-104 -223
+16 784
+771 -659
+-270 412
+-496 -688
+577 -186
+-218 -379
+-99 143
+424 490
+-282 -348
+-764 -85
+366 -621
+489 -509
+-333 -181
+501 -594
+-614 -728
+592 331
+510 195
+289 552
+-217 -743
+-665 135
+-783 714
+-318 306
+-406 -42
+342 452
+45 -502
+577 64
+646 635
+652 492
+-685 -637
+630 19
+-68 -306
+-571 148
+326 101
+206 654
+-85 -656
+-149 -30
+718 104
+441 -312
+-740 681
+-585 13
+397 224
+435 -335
+239 -432
+12 -111
+778 235
+-135 -741
+-493 -185
+-674 781
+208 -765
+-232 -789
+792 -111
+676 28
+200 59
+-503 -690
+239 -594
+-285 -187
+-292 -719
+-333 -97
+-714 -648
+671 285
+-202 -313
+-665 -241
+553 -194
+142 618
+652 521
+-400 581
+-502 715
+-530 -415
+-387 -287
+-134 528
+-176 606
+180 -120
+-327 266
+223 -44
+219 781
+-200 283
+693 70
+-764 355
+191 119
+301 -609
